deliverable-2/backend/requirements.txt
bcrypt==3.2.0
dnspython==2.0.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Ext==0.1
Flask-Login==0.5.0
flask-mongoengine==0.9.5
Flask-Mail==0.9.1
Flask-SocketIO==4.3.1
Flask-WTF==0.14.3
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
itsdangerous==1.1.0
mongoengine==0.20.0
passlib==1.7.4
PyJWT==1.7.1
Pillow==8.0.1
pymongo==3.11.0
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.24.0
urllib3==1.25.11
Werkzeug==1.0.1
wheel==0.35.1
WTForms==2.3.3
